+++ shops/views.py
@@ -1,4 +1,5 @@
 from .serializers import ShopSerializer, UpdateShopSerializer
+from django.db.models import Q
 from .models import Shop
 from accounts.models import Account
 from forums.models import Forum
@@ -28,15 +29,17 @@ class ShopListView(APIView):
     def post(self, request):
         data = request.data
         image_name = data.get('image', None)
-        image_content = data.get('image_content', None)
+        image_content = data.get('image_content', None)      
 
-        if image_content:
-            image_data = base64.b64decode(image_content)
-            with open(f'uploads/{image_name}', 'wb') as f:
-                f.write(image_data)
-        if image_name:
-            media_instance = Media.objects.create(id=image_name)
-
+        if image_name:         
+            if Media.objects.filter(Q(id=image_name)).exists():
+                return Response({"error": f"Media with name '{image_name}' already exists."}, status=status.HTTP_400_BAD_REQUEST)
+            if image_content:
+                media_instance = Media.objects.create(id=image_name)
+                image_data = base64.b64decode(image_content)
+                with open(f'uploads/{image_name}', 'wb') as f:
+                    f.write(image_data)
+        
         data.pop('image_content', None)
 
         serializer = ShopSerializer(data=data)
